ledger test tool: refactor conformance test dars
