+++ ledger-service/http-json/BUILD.bazel
@@ -153,7 +153,7 @@ da_scala_test(
     data = [
         ":Account.dar",
         "//docs:quickstart-model.dar",
-        "//ledger/test-common:Test-stable.dar",
+        "//ledger/test-common:model-tests.dar",
         "//ledger/test-common/test-certificates",
     ],
     plugins = [
