+++ ledger/ledger-api-client/BUILD.bazel
@@ -57,7 +57,7 @@ da_scala_test_suite(
     name = "ledger-api-client-integration-tests",
     srcs = glob(["src/it/**/*.scala"]),
     data = [
-        "//ledger/test-common:Test-stable.dar",
+        "//ledger/test-common:model-tests.dar",
     ],
     resources = [
         "src/it/resources/logback-test.xml",
