+++ ledger/ledger-api-test-tool/BUILD.bazel
@@ -63,7 +63,7 @@ da_scala_library(
         "//language-support/scala/bindings",
         "//ledger/ledger-api-common",
         "//ledger/test-common",
-        "//ledger/test-common:Test-stable.scala",
+        "//ledger/test-common:model-tests.scala",
         "//libs-scala/grpc-utils",
         "//libs-scala/timer-utils",
         "@maven//:ai_x_diff_2_12",
@@ -88,10 +88,9 @@ da_scala_library_suite(
         "//language-support/scala/bindings",
         "//ledger/ledger-api-common",
         "//ledger/test-common",
-        "//ledger/test-common:PerformanceTests.scala",
-        "//ledger/test-common:SemanticTests.scala",
-        "//ledger/test-common:Test-dev.scala",
-        "//ledger/test-common:Test-stable.scala",
+        "//ledger/test-common:model-tests.scala",
+        "//ledger/test-common:performance-tests.scala",
+        "//ledger/test-common:semantic-tests.scala",
         "//libs-scala/timer-utils",
         "@maven//:ai_x_diff_2_12",
         "@maven//:com_chuusai_shapeless_2_12",
@@ -154,7 +153,7 @@ conformance_test(
     ],
     test_tool_args = [
         "--verbose",
-        "--include=SemanticTests",
+        "--include=SemanticsTests",
         "--crt $(rlocation $TEST_WORKSPACE/$(rootpath //ledger/test-common/test-certificates:client.crt))",
         "--cacrt $(rlocation $TEST_WORKSPACE/$(rootpath //ledger/test-common/test-certificates:ca.crt))",
         "--pem $(rlocation $TEST_WORKSPACE/$(rootpath //ledger/test-common/test-certificates:client.pem))",
