+++ ledger/ledger-api-test-tool/src/main/scala/com/daml/ledger/api/testtool/LedgerApiTestTool.scala
@@ -92,10 +92,10 @@ object LedgerApiTestTool {
     if (config.extract) {
       // This must be kept aligned manually with artifacts declared in /ledger/test-common/BUILD.bazel.
       extractResources(
-        "/ledger/test-common/SemanticTests.dar",
-        "/ledger/test-common/Test-stable.dar",
+        "/ledger/test-common/semantic-tests.dar",
+        "/ledger/test-common/model-tests.dar",
         "/ledger/test-common/Test-dev.dar",
-        "/ledger/test-common/PerformanceTests.dar",
+        "/ledger/test-common/performance-tests.dar",
       )
       sys.exit(0)
     }
