+++ ledger/ledger-api-test-tool/src/main/scala/com/daml/ledger/api/testtool/tests/CommandServiceIT.scala
@@ -12,11 +12,11 @@ import com.daml.ledger.api.v1.commands.Command
 import com.daml.ledger.api.v1.value.{Record, RecordField, Value}
 import com.daml.ledger.client.binding.Primitive
 import com.daml.ledger.client.binding.Value.encode
-import com.daml.ledger.test_stable.Test.CallablePayout._
-import com.daml.ledger.test_stable.Test.Dummy._
-import com.daml.ledger.test_stable.Test.DummyFactory._
-import com.daml.ledger.test_stable.Test.WithObservers._
-import com.daml.ledger.test_stable.Test.{Dummy, _}
+import com.daml.ledger.test.model.Test.CallablePayout._
+import com.daml.ledger.test.model.Test.Dummy._
+import com.daml.ledger.test.model.Test.DummyFactory._
+import com.daml.ledger.test.model.Test.WithObservers._
+import com.daml.ledger.test.model.Test.{Dummy, _}
 import java.util.regex.Pattern
 import io.grpc.Status
 import scalaz.syntax.tag._
