+++ ledger/ledger-api-test-tool/src/main/scala/com/daml/ledger/api/testtool/tests/DivulgenceIT.scala
@@ -5,8 +5,8 @@ package com.daml.ledger.api.testtool.tests
 
 import com.daml.ledger.api.testtool.infrastructure.Allocation._
 import com.daml.ledger.api.testtool.infrastructure.LedgerTestSuite
-import com.daml.ledger.test_stable.Test.Divulgence2._
-import com.daml.ledger.test_stable.Test.{Divulgence1, Divulgence2}
+import com.daml.ledger.test.model.Test.Divulgence2._
+import com.daml.ledger.test.model.Test.{Divulgence1, Divulgence2}
 import scalaz.Tag
 
 final class DivulgenceIT extends LedgerTestSuite {
