+++ ledger/ledger-api-test-tool/src/main/scala/com/daml/ledger/api/testtool/tests/TransactionServiceIT.scala
@@ -17,20 +17,20 @@ import com.daml.ledger.api.v1.transaction.TreeEvent.Kind.Exercised
 import com.daml.ledger.api.v1.transaction.{Transaction, TransactionTree, TreeEvent}
 import com.daml.ledger.client.binding.Primitive
 import com.daml.ledger.client.binding.Value.encode
-import com.daml.ledger.test_stable.Iou.Iou
-import com.daml.ledger.test_stable.Iou.Iou._
-import com.daml.ledger.test_stable.Iou.IouTransfer._
-import com.daml.ledger.test_stable.IouTrade.IouTrade
-import com.daml.ledger.test_stable.IouTrade.IouTrade._
-import com.daml.ledger.test_stable.Test.Agreement._
-import com.daml.ledger.test_stable.Test.AgreementFactory._
-import com.daml.ledger.test_stable.Test.Choice1._
-import com.daml.ledger.test_stable.Test.CreateAndFetch._
-import com.daml.ledger.test_stable.Test.Dummy._
-import com.daml.ledger.test_stable.Test.DummyFactory._
-import com.daml.ledger.test_stable.Test.ParameterShowcase._
-import com.daml.ledger.test_stable.Test.TriProposal._
-import com.daml.ledger.test_stable.Test._
+import com.daml.ledger.test.model.Iou.Iou
+import com.daml.ledger.test.model.Iou.Iou._
+import com.daml.ledger.test.model.Iou.IouTransfer._
+import com.daml.ledger.test.model.IouTrade.IouTrade
+import com.daml.ledger.test.model.IouTrade.IouTrade._
+import com.daml.ledger.test.model.Test.Agreement._
+import com.daml.ledger.test.model.Test.AgreementFactory._
+import com.daml.ledger.test.model.Test.Choice1._
+import com.daml.ledger.test.model.Test.CreateAndFetch._
+import com.daml.ledger.test.model.Test.Dummy._
+import com.daml.ledger.test.model.Test.DummyFactory._
+import com.daml.ledger.test.model.Test.ParameterShowcase._
+import com.daml.ledger.test.model.Test.TriProposal._
+import com.daml.ledger.test.model.Test._
 import com.daml.platform.api.v1.event.EventOps.{EventOps, TreeEventOps}
 import io.grpc.Status
 import scalaz.Tag
